Moving node-channel tests to bolt-connection
